utilisation des intentions de l'IDE pour la migration du code vers java 7
* utilisation de collection (for -> for-each)
* visibilité de variables / méthodes
* modificateur final sur des variables

+++ GenericAHP/src/org/taeradan/ahp/ConsistencyChecker.java
@@ -162,8 +162,8 @@ public final class ConsistencyChecker {
 		assert consistencyData.getConsistencyRatio().isInfinite() == false;
 		assert consistencyData.getConsistencyRatio().isNaN() == false;
 
-		assert consistencyData.getConsistencyRatio().doubleValue() <= 100.0 / 100.0;
-		assert consistencyData.getConsistencyRatio().doubleValue() >= 0.0 / 100.0;
+		assert consistencyData.getConsistencyRatio() <= 100.0 / 100.0;
+		assert consistencyData.getConsistencyRatio() >= 0.0 / 100.0;
 
 		return consistencyData.getConsistencyRatio();
 	}
